Ooc rule CRUD
