+++ zing_product_backend/services/general_settings/crud.py
@@ -8,6 +8,12 @@ from zing_product_backend.models import general_settings
 from zing_product_backend.core import common
 from zing_product_backend.services.general_settings import util_functions
 from zing_product_backend.models.auth_model import User
+from zing_product_backend.models.containment_model import ContainmentRule
+from zing_product_backend.models.general_settings import OOCRules, ContainmentRule
+from datetime import datetime
+from zing_product_backend.core.security.schema import UserInfo
+from sqlalchemy import select, update, delete
+
 
 class DatabaseError(Exception):
     pass
@@ -335,3 +341,65 @@ class SettingsDataBase:
             await self.session.delete(group_orm)
         await self.session.commit()
         return deleted_group_orm
+
+
+class OOCRulesCRUD:
+    def __init__(self, session: AsyncSession):
+        self.session = session
+
+    async def create_ooc_rule(self, ooc_rule_data: schemas.OOCRuleCreate, user: UserInfo) -> OOCRules:
+        new_ooc_rule = OOCRules(
+            containment_rule_id=ooc_rule_data.containment_rule_id,
+            spec_id=ooc_rule_data.spec_id,
+            lower_limit=ooc_rule_data.lower_limit,
+            upper_limit=ooc_rule_data.upper_limit,
+            create_user_name=user.user_name,
+            updated_user_name=user.user_name,
+            create_time=datetime.now(),
+            updated_time=datetime.now(),
+            rule_delete_flag=False
+        )
+        self.session.add(new_ooc_rule)
+        await self.session.commit()
+        await self.session.refresh(new_ooc_rule)
+        return new_ooc_rule
+
+    async def update_ooc_rule(self, update_data: schemas.OOCRuleUpdate, user: UserInfo) -> None:
+        update_stmt = update(OOCRules).where(OOCRules.id == update_data.id).values(
+            lower_limit=update_data.lower_limit,
+            upper_limit=update_data.upper_limit,
+            updated_time=datetime.now(),
+            updated_user_name=user.user_name
+        )
+        await self.session.execute(update_stmt)
+        await self.session.commit()
+
+    async def delete_ooc_rule(self, ooc_rule_id: int, user: UserInfo) -> None:
+        stmt = select(OOCRules).where(OOCRules.id == ooc_rule_id)
+        ooc_orm = (await self.session.execute(stmt)).scalars().first()
+        if ooc_orm is None:
+            raise DatabaseError(f'mat_id: {ooc_rule_id} not found')
+        delete_stmt = update(OOCRules).where(OOCRules.id == ooc_rule_id).values(
+            rule_delete_flag=True,
+            updated_time=datetime.now(),
+            updated_user_name=user.user_name
+        )
+        await self.session.execute(delete_stmt)
+        await self.session.commit()
+
+    async def get_ooc_rule_by_id(self, ooc_rule_id: int) -> OOCRules:
+        select_stmt = select(OOCRules).where(and_(OOCRules.id == ooc_rule_id, OOCRules.rule_delete_flag == False))
+        result = (await self.session.execute(select_stmt)).scalars().first()
+        return result
+
+    async def get_all_ooc_rules(self) -> list[OOCRules]:
+        select_stmt = select(OOCRules).where(OOCRules.rule_delete_flag == False)
+        result = await self.session.execute(select_stmt)
+        return result.scalars().all()
+
+    async def get_ooc_rule_by_name(self, ooc_rule_name: str) -> OOCRules:
+        info = select(ContainmentRule.id).where(ContainmentRule.rule_name == ooc_rule_name)
+        select_stmt = select(OOCRules).where(and_(OOCRules.containment_rule_id == info, OOCRules.rule_delete_flag == False))
+        print(select_stmt)
+        result = await self.session.execute(select_stmt)
+        return result.scalars().all()
